Restructure seviri code
